chore: 코드 폴더 분리 (#60)
