+++ frontend/src/component/main/intro/trimCard/trimDescription/TrimDescription.tsx
