+++ frontend/src/component/main/intro/trimCard/trimDescription/trimDescriptionDetail/DetailDescription.tsx
