+++ frontend/src/component/main/intro/trimCard/trimDescription/trimDescriptionDetail/DetailFeature.tsx
