+++ frontend/src/component/main/intro/trimCard/trimDescription/trimDescriptionDetail/TrimDescriptionDetail.tsx
